+++ src/bda/plone/shop/interfaces.py
@@ -12,6 +12,7 @@ from zope import schema
 from zope.interface import Attribute
 from zope.interface import Interface
 from zope.interface import provider
+
 import zope.deferredimport