+++ src/bda/plone/shop/vocabularies.py
@@ -2,7 +2,7 @@
 from bda.plone.checkout.vocabularies import country_vocabulary
 from bda.plone.checkout.vocabularies import gender_vocabulary
 from bda.plone.payment import Payments
-from bda.plone.shipping import Shippings
+from bda.plone.cart.shipping import Shippings
 from bda.plone.shop import message_factory as _
 from bda.plone.shop.utils import get_shop_article_settings
 from bda.plone.shop.utils import get_shop_tax_settings